fix(remoter): 自动滚动到底部
